PEAR-1983 - add other clinical attributes, remove deprecated attributes, break up diagnoses and follow ups
