+++ packages/portal-proto/src/features/cases/ClinicalSummary/tests/ClinicalSummary.unit.test.tsx
@@ -1,10 +1,7 @@
 import { render } from "test-utils";
 import { ClinicalSummary } from "../ClinicalSummary";
 import userEvent from "@testing-library/user-event";
-import {
-  mockSingleDiagnoses,
-  mockSingleFollowUps,
-} from "./DiagnosesOrFollowUps.unit.test";
+import { mockSingleDiagnoses, mockSingleFollowUps } from "./mockData";
 import {
   mock_single_exposures,
   mock_single_family_histories,
