+++ packages/portal-proto/src/features/cases/utils.tsx
@@ -75,18 +75,12 @@ export const caseSummaryFields = [
   "follow_ups.follow_up_id",
   "follow_ups.submitter_id",
   "follow_ups.days_to_follow_up",
-  "follow_ups.comorbidity",
-  "follow_ups.risk_factor",
   "follow_ups.progression_or_recurrence_type",
   "follow_ups.progression_or_recurrence",
   "follow_ups.disease_response",
-  "follow_ups.bmi",
-  "follow_ups.height",
-  "follow_ups.weight",
   "follow_ups.ecog_performance_status",
   "follow_ups.karnofsky_performance_status",
   "follow_ups.progression_or_recurrence_anatomic_site",
-  "follow_ups.reflux_treatment_type",
   "follow_ups.molecular_tests.aa_change",
   "follow_ups.molecular_tests.antigen",
   "follow_ups.molecular_tests.biospecimen_type",
@@ -103,6 +97,14 @@ export const caseSummaryFields = [
   "follow_ups.molecular_tests.test_units",
   "follow_ups.molecular_tests.test_value",
   "follow_ups.molecular_tests.variant_type",
+  "follow_ups.other_clinical_attributes.submitter_id",
+  "follow_ups.other_clinical_attributes.other_clinical_attribute_id",
+  "follow_ups.other_clinical_attributes.timepoint_category",
+  "follow_ups.other_clinical_attributes.nononcologic_therapeutic_agents",
+  "follow_ups.other_clinical_attributes.treatment_frequency",
+  "follow_ups.other_clinical_attributes.weight",
+  "follow_ups.other_clinical_attributes.height",
+  "follow_ups.other_clinical_attributes.bmi",
 ];
 
 export const getSlideCountFromCaseSummary = (
